Unify unix config parser
