+++ unix_integration/common/Cargo.toml
@@ -14,6 +14,8 @@ repository = { workspace = true }
 [features]
 default = ["unix"]
 unix = []
+selinux = []
+tpm = []
 
 [lib]
 name = "kanidm_unix_common"
